Use dynamic port mapping for the interactive script
I go back and forth on this. It's nice to not have any port conflicts if
you just want to pull up `--help` in another terminal, and there are too
many complexities with <wikifolder> path handling to make using the
interactive script realistic...the `tiddlywiki-server` script is much
better for that.