Namespace the debugging environment variable
For consistency and to prevent conflicts, we should prefix all of the
environment variables that are specific to tiddlywiki-docker with
`TIDDLYWIKI_DOCKER_`. It's a bit long, but shouldn't be needed too
often, has an obvious meaning, and mirrors both what TiddlyWiki itself
uses as well as what we expose within the container as metadata.