Add first attempt at a Circle CI configuration
Since we've configured automated builds with Docker Hub to tag both
`latest` and the major version number (e.g. `5`) from the _master_
branch, the idea here is that if tests pass on the master branch, we'll
go ahead and alias the `5` tag with the version-specific tag, like
`5.1.17`.

I previously had Docker Hub building on Git tags for those, but I'd like
to be able to rebuild the most recent tags on a somewhat regular basis
to pick up any system-level updates (for security and whatnot). Changing
Git tags is problematic since people have to explicitly delete then and
re-sync (it won't happen transparently)...so we'll try doing it all at
the registry level.

This might take some re-working once there's actually a new major
version, but I just want to get something working to start with so
I don't have to manually push tags.