Add patch version tag to major version image on daily basis
There's no reason to push the actual image we build as part of testing
when Docker Hub is already building the major version images for us
through the build automation feature. I'd rather the one image that
Docker Hub creates have multiple tags associated with it; but the actual
patch version tagging should be automated.

When a new commit is pushed to the master branch, there could
potentially be a race condition between Docker Hub building the new
images and Circle CI testing the master branch, so I don't want to
blindly pull the major and tag it as a patch version on every commit.

Instead, I'm attempting to use workflows to make the distinction between
the testing job that happens on pushes, and the tagging job that happens
on a daily basis via a scheduled trigger.

Until I can sort out the kinks, the actual push command is not yet
running.