Manage latest/patch/major version tags via CI
I realized that with the Docker Hub automated builds, it's possible that
an untested image could be pushed. I'd rather manage all the important
tags via CI so I can only push tested images, and also the same image
can be used for `latest` as well as the most recent major version and
patch version.

I still have the build automation enabled on Docker Hub, but am pushing
a tag named `readme` that I hope will be ignored. I do that so the
README is automatically updated, and the latest _Dockerfile_ will also
be displayed.

I've also removed any version number from the CI configuration, and
instead added a symlink at the top-level of the repo to point to the
latest version. I think this should work long-term and I can always
rebuild certain versions manually and push things if necessary.