Add a Makefile to handle common project commands
This allows us to separate out the commands that are useful for
development and/or manual management of images from the commands that
are specific to running within a CI environment.

I'm not sure what the best conventions are regarding the naming for tags
used when developing locally, but I'm sticking with a bare `tiddlywiki`
with no repository name. That allows the actual `DOCKER_HUB_REPOSITORY`
value to be abstracted out into a single variable (that's set by default
to `elasticdog`), which is only really used for the deploy step which
pushes it to the Docker Hub registry.