Remove container health output from JUnit XML report
This is pushed to STDOUT by `goss`, and I couldn't find a way to disable
that particular check; so in order for us to have a clean XML file, we
must remove that output.

The `sed` command uses `-n` to not print any lines by default, it
matches the opening XML tag, and then prints from there to end of the
file.