Add automated monthly maintenance builds
This will allow for picking up any base image updates (Alpine and/or
Node.js) for security purposes even if there are no new commits in the
tiddlywiki-docker repository. Most of the official distros only push new
tags about once a month, and I don't want to be continually pestered if
there's a breaking change, so a monthly schedule seemed reasonable.

For people that need consistency for deployments, there are always the
static unique tags.