Always make a unique tag, locally or in CI
The reason for supporting a local unique tag would be if a manual build
happens for some reason and needs to get pushed. That said, when running
locally, we use the short Git commit SHA instead of the build number,
since that's not applicable. It's a sane fallback, but one that
shouldn't come up very often.